Updated Login modal to use redux state
Updated login modal to use redux state (so that it can be opened and closed from anywhere as needed). The particular use case involved not closing the login modal until the user had successfully logged in (or chosen to close the modal). 

Also included some minor bug fixes. Namely, a fix to ensure cards and notes were loaded when the user logged in.