Update to fix issue with private route
Fixed issue on Study Deck page where entire component would remount when a note was saved. Issue was related to React Router rendering a component that was sent via the private route. This was resolved by changing react router to use a sent render function instead of a component to render.